Validation of stack/version tag added. Improved error checking for SplitVersionFromStack. util functions moved to util.go
